docs/requirements.txt
furo == 2022.3.4
sphinx-copybutton == 0.5.0
sphinx-inline-tabs == 2022.1.2b11
